+++ src/mutate_and_summ.jl
@@ -112,6 +112,14 @@ macro mutate(sqlquery, mutations...)
 
             if sq.post_aggregation #|| sq.post_join 
                 # Reset post_aggregation as we're now handling it
+
+                if sq.post_aggregation
+                    for row in eachrow(sq.metadata)
+                        if row[:current_selxn] == 2
+                            row[:current_selxn] = 1
+                        end
+                    end
+                end
                 sq.post_aggregation = false
                
                 select_expressions = !isempty(sq.select) ? [sq.select] : ["*"]
@@ -132,6 +140,8 @@ macro mutate(sqlquery, mutations...)
 
                 # Create and add the new CTE
                 new_cte = CTE(name=string(cte_name), select=cte_sql)
+                up_cte_name(sq, string(cte_name))
+                
                 push!(sq.ctes, new_cte)
                 sq.cte_count += 1
                 sq.from = string(cte_name)
@@ -139,27 +149,21 @@ macro mutate(sqlquery, mutations...)
             else
                 sq.cte_count += 1
                 cte_name = "cte_" * string(sq.cte_count)
+                
             end
-
             cte_name = "cte_" * string(sq.cte_count + 1)
             sq.cte_count += 1
 
             # Prepare select expressions, starting with existing selections if any
+           
             select_expressions = ["*"]
+
             all_columns = [
                 (row[:current_selxn] == 1 ? row[:name] : row[:table_name] * "." * row[:name])
                 for row in eachrow(sq.metadata) if row[:current_selxn] != 0
             ]            
             select_expressions = [col for col in all_columns]  # Start with all currently selected columns
-            # Set the grouping variable if `by` is provided
-            #println("here" ,select_expressions, sq.post_join)
-            if !sq.post_join
-                select_expressions = [
-                    occursin(".", expr) ? join(split(expr, ".")[2:end], ".") : expr 
-                    for expr in select_expressions
-                ]
-             #   println(select_expressions)
-            end
+
             if $(esc(grouping_var)) != nothing
                 group_vars = $(esc(grouping_var))
                 group_vars_sql = expr_to_sql(group_vars, sq)
@@ -205,10 +209,12 @@ macro mutate(sqlquery, mutations...)
 
             # Create and add the new CTE
             new_cte = CTE(name=string(cte_name), select=cte_sql)
+            up_cte_name(sq, cte_name)
             push!(sq.ctes, new_cte)
 
-            # Update sq.from to the latest CTE, reset sq.select for final query
+
             sq.from = string(cte_name)
+
             sq.select = "*"
             if _warning_[]
                 if sq.groupBy != "" || sq.window_order != "" || sq.windowFrame != ""
@@ -334,6 +340,7 @@ macro summarize(sqlquery, expressions...)
 
             sq.is_aggregated = true        # Mark the query as aggregated
             sq.post_aggregation = true     # Indicate ready for post-aggregation operations
+            
         else
             error("Expected sqlquery to be an instance of SQLQuery") # COV_EXCL_LINE
         end