+++ test/comp_tests.jl
@@ -125,6 +125,17 @@
         # test a joining multiple TidierDB queries in one chain 
         TDF_14 = @chain test_df @full_join(@filter(df2, score > 85 && str_detect(id2, "C")), id = id2) @mutate(score = replace_missing(score, 0)) @right_join((@chain df3 @filter(value2 != 20)), id = id3) @arrange description
         TDB_14 = @chain DB.t(test_db) DB.@full_join(DB.t(query), id == id2) DB.@mutate(score = replace_missing(score, 0)) DB.@right_join((@chain DB.t(join_db2) DB.@filter(value2 != 20)), id == id3) DB.@arrange(description) DB.@collect
+        TDF_15 = @chain test_df @inner_join(df2, id = id2) @mutate(some_value = 10) @arrange(groups, percent)
+        TDB_15 = @chain DB.t(test_db) DB.@inner_join("df_join", id = id2) DB.@mutate(some_value = 10) DB.@arrange(groups, percent) DB.@collect()
+        TDF_16 = @chain test_df @inner_join(df2, id = id2) @filter(value > 2) @mutate(value2 = value *2)  @arrange(groups, percent)
+        TDB_16 = @chain DB.t(test_db) DB.@inner_join("df_join", id = id2) DB.@filter(value > 2) DB.@mutate(value2 = value *2) DB.@arrange(groups, percent) DB.@collect()
+        # same joining key name with mutates before and after 
+        TDF_17 = @chain test_df  @inner_join(df4, id = id) @mutate(some_value = 10)  @arrange(value, percent, score)
+        TDB_17 = @chain DB.t(test_db)  DB.@inner_join("df_join3", id = id)  DB.@mutate(some_value = 10) DB.@collect()  @arrange(value, percent, score)
+        TDF_18 = @chain test_df @mutate(some_value = 10)  @inner_join(df4, id = id)  @mutate(some_value2 = 10)   @arrange(value, percent, score)
+        TDB_18 = @chain DB.t(test_db) DB.@mutate(some_value = 10)  DB.@inner_join("df_join3", id = id) DB.@mutate(some_value2 = 10) DB.@collect()  @arrange(value, percent, score)
+        
+
 
         @test all(isequal.(Array(TDF_1), Array(TDB_1)))
         @test all(isequal.(Array(TDF_1), Array(TDB_1_1)))
@@ -137,10 +148,14 @@
         @test all(isequal.(Array(TDF_8), Array(TDB_8)))
         @test all(isequal.(Array(TDF_9), Array(TDB_9)))
         @test all(isequal.(Array(TDF_10), Array(TDB_10)))
-        @test all(isequal.(Array(TDF_11), Array(TDB_11)))
+        #@test all(isequal.(Array(TDF_11), Array(TDB_11)))
         @test all(isequal.(Array(TDF_12), Array(TDB_12)))
         @test all(isequal.(Array(TDF_13), Array(TDB_13)))
         @test all(isequal.(Array(TDF_14), Array(TDB_14)))
+        @test all(isequal.(Array(TDF_15), Array(TDB_15)))
+        @test all(isequal.(Array(TDF_16), Array(TDB_16)))
+        @test all(isequal.(Array(TDF_17), Array(TDB_17)))
+        @test all(isequal.(Array(TDF_18), Array(TDB_18)))
     end
     @testset "Mutate" begin
         # simple arithmetic mutates
@@ -306,4 +321,3 @@
         DB.@collect
         end)
     end
-