+++ NEWS.md
@@ -5,6 +5,7 @@
     - `lead`, `lag`, `dense_rank`, `nth_value`, `ntile`, `rank_dense`, `row_number`, `first_value`, `last_value`, `cume_dist`
 - add ability to change what functions are on this list to avoid the use of agg in the following manner 
     - `push!(TidierDB.window_agg_fxns, :kurtosis);`
+- fixes edge case query construction issues with `@mutate`,`@filter`,`@*_join`
 
 ## v0.7.0 - 2025-01-26
 - `db_table` now supports viewing a dataframe directly - `db_table(db, df, "name4db")`